refactor for footprint
