+++ src/main/java/com/alibaba/fastjson/parser/deserializer/OptionalCodec.java
@@ -7,13 +7,12 @@ import java.util.OptionalDouble;
 import java.util.OptionalInt;
 import java.util.OptionalLong;
 
+import com.alibaba.fastjson.JSONException;
 import com.alibaba.fastjson.parser.DefaultJSONParser;
 import com.alibaba.fastjson.parser.JSONToken;
 import com.alibaba.fastjson.serializer.JSONSerializer;
 import com.alibaba.fastjson.serializer.ObjectSerializer;
 import com.alibaba.fastjson.util.TypeUtils;
-
-import net.sf.json.JSONException;
 
 public class OptionalCodec implements ObjectSerializer, ObjectDeserializer {
