+++ src/main/java/com/alibaba/fastjson/parser/deserializer/SqlDateDeserializer.java
@@ -1,6 +1,7 @@
 package com.alibaba.fastjson.parser.deserializer;
 
 import java.lang.reflect.Type;
+import java.sql.Timestamp;
 import java.text.DateFormat;
 import java.text.ParseException;
 import java.util.Date;
@@ -13,9 +14,24 @@ import com.alibaba.fastjson.parser.JSONToken;
 public class SqlDateDeserializer extends AbstractDateDeserializer implements ObjectDeserializer {
 
     public final static SqlDateDeserializer instance = new SqlDateDeserializer();
+    public final static SqlDateDeserializer instance_timestamp = new SqlDateDeserializer(true);
+    
+    private boolean                           timestamp = false;
+    
+    public SqlDateDeserializer() {
+        
+    }
+    
+    public SqlDateDeserializer(boolean timestmap) {
+        this.timestamp = true;
+    }
 
     @SuppressWarnings("unchecked")
     protected <T> T cast(DefaultJSONParser parser, Type clazz, Object fieldName, Object val) {
+        if (timestamp) {
+            return castTimestamp(parser, clazz, fieldName, val);
+        }
+        
         if (val == null) {
             return null;
         }
@@ -58,6 +74,42 @@ public class SqlDateDeserializer extends AbstractDateDeserializer implements Obj
         }
 
         return (T) val;
+    }
+    
+    @SuppressWarnings("unchecked")
+    protected <T> T castTimestamp(DefaultJSONParser parser, Type clazz, Object fieldName, Object val) {
+
+        if (val == null) {
+            return null;
+        }
+
+        if (val instanceof java.util.Date) {
+            return (T) new java.sql.Timestamp(((Date) val).getTime());
+        }
+
+        if (val instanceof Number) {
+            return (T) new java.sql.Timestamp(((Number) val).longValue());
+        }
+
+        if (val instanceof String) {
+            String strVal = (String) val;
+            if (strVal.length() == 0) {
+                return null;
+            }
+
+            DateFormat dateFormat = parser.getDateFormat();
+            try {
+                Date date = (Date) dateFormat.parse(strVal);
+                return (T) new Timestamp(date.getTime());
+            } catch (ParseException e) {
+                // skip
+            }
+
+            long longVal = Long.parseLong(strVal);
+            return (T) new java.sql.Timestamp(longVal);
+        }
+
+        throw new JSONException("parse error");
     }
 
     public int getFastMatchToken() {
