--- src/main/java/com/alibaba/fastjson/serializer/AtomicLongArrayCodec.java
@@ -1,80 +1,0 @@
-/*
- * Copyright 1999-2101 Alibaba Group.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.alibaba.fastjson.serializer;
-
-import java.io.IOException;
-import java.lang.reflect.Type;
-import java.util.concurrent.atomic.AtomicLongArray;
-
-import com.alibaba.fastjson.JSONArray;
-import com.alibaba.fastjson.parser.DefaultJSONParser;
-import com.alibaba.fastjson.parser.JSONToken;
-import com.alibaba.fastjson.parser.deserializer.ObjectDeserializer;
-
-/**
- * @author wenshao[szujobs@hotmail.com]
- */
-public class AtomicLongArrayCodec implements ObjectSerializer, ObjectDeserializer {
-
-    public final static AtomicLongArrayCodec instance = new AtomicLongArrayCodec();
-
-    public void write(JSONSerializer serializer, Object object, Object fieldName, Type fieldType, int features) throws IOException {
-        SerializeWriter out = serializer.getWriter();
-
-        if (object == null) {
-            if (out.isEnabled(SerializerFeature.WriteNullListAsEmpty)) {
-                out.write("[]");
-            } else {
-                out.writeNull();
-            }
-            return;
-        }
-
-        AtomicLongArray array = (AtomicLongArray) object;
-        int len = array.length();
-        out.append('[');
-        for (int i = 0; i < len; ++i) {
-            long val = array.get(i);
-            if (i != 0) {
-                out.write(',');
-            }
-            out.writeLong(val);
-        }
-        out.append(']');
-    }
-
-    @SuppressWarnings("unchecked")
-    public <T> T deserialze(DefaultJSONParser parser, Type clazz, Object fieldName) {
-        if (parser.getLexer().token() == JSONToken.NULL) {
-            parser.getLexer().nextToken(JSONToken.COMMA);
-            return null;
-        }
-
-        JSONArray array = new JSONArray();
-        parser.parseArray(array);
-
-        AtomicLongArray atomicArray = new AtomicLongArray(array.size());
-        for (int i = 0; i < array.size(); ++i) {
-            atomicArray.set(i, array.getLong(i));
-        }
-
-        return (T) atomicArray;
-    }
-
-    public int getFastMatchToken() {
-        return JSONToken.LBRACKET;
-    }
-}
