--- src/main/java/com/alibaba/fastjson/serializer/AtomicLongSerializer.java
@@ -1,36 +1,0 @@
-/*
- * Copyright 1999-2101 Alibaba Group.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.alibaba.fastjson.serializer;
-
-import java.io.IOException;
-import java.lang.reflect.Type;
-import java.util.concurrent.atomic.AtomicLong;
-
-/**
- * @author wenshao[szujobs@hotmail.com]
- */
-public class AtomicLongSerializer implements ObjectSerializer {
-
-    public final static AtomicLongSerializer instance = new AtomicLongSerializer();
-
-    public void write(JSONSerializer serializer, Object object, Object fieldName, Type fieldType, int features) throws IOException {
-        SerializeWriter out = serializer.getWriter();
-
-        AtomicLong val = (AtomicLong) object;
-        out.writeLong(val.get());
-    }
-
-}
