+++ src/main/java/com/alibaba/fastjson/serializer/CharArrayCodec.java
@@ -1,15 +1,34 @@
-package com.alibaba.fastjson.parser.deserializer;
+package com.alibaba.fastjson.serializer;
 
+import java.io.IOException;
 import java.lang.reflect.Type;
 
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.parser.DefaultJSONParser;
 import com.alibaba.fastjson.parser.JSONLexer;
 import com.alibaba.fastjson.parser.JSONToken;
+import com.alibaba.fastjson.parser.deserializer.ObjectDeserializer;
 
-public class CharArrayDeserializer implements ObjectDeserializer {
 
-    public final static CharArrayDeserializer instance = new CharArrayDeserializer();
+public class CharArrayCodec implements ObjectSerializer, ObjectDeserializer {
+
+    public static CharArrayCodec instance = new CharArrayCodec();
+
+    public final void write(JSONSerializer serializer, Object object, Object fieldName, Type fieldType, int features) throws IOException {
+        SerializeWriter out = serializer.getWriter();
+        
+        if (object == null) {
+            if (out.isEnabled(SerializerFeature.WriteNullListAsEmpty)) {
+                out.write("[]");
+            } else {
+                out.writeNull();
+            }
+            return;
+        }
+
+        char[] chars = (char[]) object;
+        out.writeString(new String(chars));
+    }
 
     @SuppressWarnings("unchecked")
     public <T> T deserialze(DefaultJSONParser parser, Type clazz, Object fieldName) {
@@ -43,5 +62,4 @@ public class CharArrayDeserializer implements ObjectDeserializer {
     public int getFastMatchToken() {
         return JSONToken.LITERAL_STRING;
     }
-
 }
