+++ src/main/java/com/alibaba/fastjson/serializer/ClassCodec.java
@@ -1,0 +1,64 @@
+/*
+ * Copyright 1999-2101 Alibaba Group.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.alibaba.fastjson.serializer;
+
+import java.io.IOException;
+import java.lang.reflect.Type;
+
+import com.alibaba.fastjson.JSONException;
+import com.alibaba.fastjson.parser.DefaultJSONParser;
+import com.alibaba.fastjson.parser.JSONLexer;
+import com.alibaba.fastjson.parser.JSONToken;
+import com.alibaba.fastjson.parser.deserializer.ObjectDeserializer;
+import com.alibaba.fastjson.util.TypeUtils;
+
+/**
+ * @author wenshao[szujobs@hotmail.com]
+ */
+public class ClassCodec implements ObjectSerializer, ObjectDeserializer {
+
+    public final static ClassCodec instance = new ClassCodec();
+
+    @SuppressWarnings("rawtypes")
+    public void write(JSONSerializer serializer, Object object, Object fieldName, Type fieldType, int features) throws IOException {
+        SerializeWriter out = serializer.getWriter();
+
+        Class clazz = (Class) object;
+        out.writeString(clazz.getName());
+    }
+
+    @SuppressWarnings("unchecked")
+    public <T> T deserialze(DefaultJSONParser parser, Type type, Object fieldName) {
+        JSONLexer lexer = parser.getLexer();
+        
+        if (lexer.token() == JSONToken.NULL) {
+            lexer.nextToken();
+            return null;
+        }
+        
+        if (lexer.token() != JSONToken.LITERAL_STRING) {
+            throw new JSONException("expect className");
+        }
+        String className = lexer.stringVal();
+        lexer.nextToken(JSONToken.COMMA);
+
+        return (T) TypeUtils.loadClass(className, parser.getConfig().getDefaultClassLoader());
+    }
+
+    public int getFastMatchToken() {
+        return JSONToken.LITERAL_STRING;
+    }
+}
