--- src/main/java/com/alibaba/fastjson/serializer/ObjectArraySerializer.java
@@ -1,119 +1,0 @@
-/*
- * Copyright 1999-2101 Alibaba Group.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.alibaba.fastjson.serializer;
-
-import java.io.IOException;
-import java.lang.reflect.Type;
-
-/**
- * @author wenshao[szujobs@hotmail.com]
- */
-public class ObjectArraySerializer implements ObjectSerializer {
-
-    public static final ObjectArraySerializer instance = new ObjectArraySerializer();
-
-    public ObjectArraySerializer(){
-    }
-
-    public final void write(JSONSerializer serializer, Object object, Object fieldName, Type fieldType, int features)
-                                                                                                       throws IOException {
-        SerializeWriter out = serializer.getWriter();
-
-        Object[] array = (Object[]) object;
-
-        if (object == null) {
-            if (out.isEnabled(SerializerFeature.WriteNullListAsEmpty)) {
-                out.write("[]");
-            } else {
-                out.writeNull();
-            }
-            return;
-        }
-
-        int size = array.length;
-
-        int end = size - 1;
-
-        if (end == -1) {
-            out.append("[]");
-            return;
-        }
-
-        SerialContext context = serializer.getContext();
-        serializer.setContext(context, object, fieldName, 0);
-
-        try {
-            Class<?> preClazz = null;
-            ObjectSerializer preWriter = null;
-            out.append('[');
-
-            if (out.isEnabled(SerializerFeature.PrettyFormat)) {
-                serializer.incrementIndent();
-                serializer.println();
-                for (int i = 0; i < size; ++i) {
-                    if (i != 0) {
-                        out.write(',');
-                        serializer.println();
-                    }
-                    serializer.write(array[i]);
-                }
-                serializer.decrementIdent();
-                serializer.println();
-                out.write(']');
-                return;
-            }
-
-            for (int i = 0; i < end; ++i) {
-                Object item = array[i];
-
-                if (item == null) {
-                    out.append("null,");
-                } else {
-                    if (serializer.containsReference(item)) {
-                        serializer.writeReference(item);
-                    } else {
-                        Class<?> clazz = item.getClass();
-
-                        if (clazz == preClazz) {
-                            preWriter.write(serializer, item, null, null, 0);
-                        } else {
-                            preClazz = clazz;
-                            preWriter = serializer.getObjectWriter(clazz);
-
-                            preWriter.write(serializer, item, null, null, 0);
-                        }
-                    }
-                    out.append(',');
-                }
-            }
-
-            Object item = array[end];
-
-            if (item == null) {
-                out.append("null]");
-            } else {
-                if (serializer.containsReference(item)) {
-                    serializer.writeReference(item);
-                } else {
-                    serializer.writeWithFieldName(item, end);
-                }
-                out.append(']');
-            }
-        } finally {
-            serializer.setContext(context);
-        }
-    }
-}
